Prompt to overwrite an existing file (#1)
The current behavior is to terminate if the output file already exists.  This adds friction, especially since the output files are intended to be checked in to version control.  This adds a prompt to let the user choose whether to continue or to terminate, along with a `--yes` option to suppress the prompt entirely.

This also fixes a bug where "info" lines on the console are not printed in the default text color after printing a colored line.